refactor ValidationOverview.tsx into 'submitProposal', changes to fetch cycle dates and the select cycle input
