+++ src/main/webui/src/ProposalEditorView/submitProposal/ValidationOverview.tsx
